+++ kitsune-cli/Cargo.toml
@@ -9,7 +9,7 @@ clap = { version = "4.1.4", features = ["derive"] }
 dotenvy = "0.15.6"
 envy = "0.4.2"
 kitsune-db = { version = "0.1.0", path = "../kitsune-db" }
-sea-orm = { version = "0.10.7", features = [
+sea-orm = { version = "0.11.0", features = [
     "sqlx-postgres",
     "runtime-tokio-rustls",
 ] }
