+++ kitsune-db/migration/Cargo.toml
@@ -12,7 +12,7 @@ path = "src/lib.rs"
 async-std = { version = "1.12.0", features = ["attributes", "tokio1"] }
 
 [dependencies.sea-orm-migration]
-version = "0.10.7"
+version = "0.11.0"
 features = [
   # Enable at least one `ASYNC_RUNTIME` and `DATABASE_DRIVER` feature if you want to run migration via CLI.
   # View the list of supported features at https://www.sea-ql.org/SeaORM/docs/install-and-config/database-and-async-runtime.
