+++ kitsune-http-client/Cargo.toml
@@ -16,7 +16,7 @@ hyper = { version = "0.14.24", features = [
 hyper-rustls = { version = "0.23.2", features = ["http2"] }
 kitsune-http-signatures = { path = "../kitsune-http-signatures" }
 serde = "1.0.152"
-serde_json = "1.0.91"
+serde_json = "1.0.92"
 tower = "0.4.13"
 tower-http = { version = "0.3.5", features = [
     "decompression-full",
