+++ kitsune-type/Cargo.toml
@@ -6,5 +6,8 @@ edition = "2021"
 [dependencies]
 chrono = { version = "0.4.23", features = ["serde"] }
 serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.91"
+serde_json = "1.0.92"
 uuid = { version = "1.3.0", features = ["serde"] }
+
+[dev-dependencies]
+pretty_assertions = "1.3.0"
