+++ kitsune-type/src/nodeinfo/two_one.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum Protocol {
     ActivityPub,
@@ -16,7 +16,7 @@ pub enum Protocol {
     Zot,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum InboundService {
     #[serde(rename = "atom1.0")]
@@ -31,7 +31,7 @@ pub enum InboundService {
     Twitter,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum OutboundService {
     #[serde(rename = "atom1.0")]
@@ -66,45 +66,45 @@ pub enum OutboundService {
     Xmpp,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 pub enum Version {
     #[serde(rename = "2.1")]
     TwoOne,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Software {
     pub name: String,
     pub version: String,
     pub repository: String,
-    pub homepage: String,
+    pub homepage: Option<String>,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Services {
     pub inbound: Vec<InboundService>,
     pub outbound: Vec<OutboundService>,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct UsageUsers {
     pub total: u64,
-    pub active_halfyear: u64,
-    pub active_month: u64,
+    pub active_halfyear: Option<u64>,
+    pub active_month: Option<u64>,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Usage {
     pub users: UsageUsers,
     pub local_posts: u64,
-    pub local_comments: u64,
+    pub local_comments: Option<u64>,
 }
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Eq, PartialEq, Serialize)]
 #[serde(rename_all = "camelCase")]
 /// Definitions of Nodeinfo 2.1
 pub struct TwoOne {
@@ -115,4 +115,20 @@ pub struct TwoOne {
     pub open_registrations: bool,
     pub usage: Usage,
     pub metadata: Value,
+}
+
+#[cfg(test)]
+mod test {
+    use super::TwoOne;
+    use crate::nodeinfo::two_one::Protocol;
+    use pretty_assertions::{assert_eq, assert_str_eq};
+
+    #[test]
+    fn deserialize_akkoma_nodeinfo() {
+        let raw = include_str!("../../tests/nodeinfo_2.1.json");
+        let two_one: TwoOne = serde_json::from_str(raw).unwrap();
+
+        assert_str_eq!(two_one.software.name, "akkoma");
+        assert_eq!(two_one.protocols, [Protocol::ActivityPub]);
+    }
 }
