+++ kitsune-type/tests/nodeinfo_2.1.json
@@ -1,0 +1,160 @@
+{
+    "metadata": {
+        "accountActivationRequired": false,
+        "features": [
+            "pleroma_api",
+            "akkoma_api",
+            "mastodon_api",
+            "mastodon_api_streaming",
+            "polls",
+            "v2_suggestions",
+            "pleroma_explicit_addressing",
+            "shareable_emoji_packs",
+            "multifetch",
+            "pleroma:api/v1/notifications:include_types_filter",
+            "editing",
+            "media_proxy",
+            "pleroma_emoji_reactions",
+            "exposable_reactions",
+            "profile_directory",
+            "custom_emoji_reactions",
+            "pleroma:get:main/ostatus"
+        ],
+        "federation": {
+            "enabled": true,
+            "exclusions": false,
+            "mrf_hashtag": {
+                "federated_timeline_removal": [],
+                "reject": [],
+                "sensitive": [
+                    "nsfw"
+                ]
+            },
+            "mrf_policies": [
+                "SimplePolicy",
+                "TagPolicy",
+                "InlineQuotePolicy",
+                "HashtagPolicy",
+                "NormalizeMarkup"
+            ],
+            "mrf_simple": {
+                "accept": [],
+                "avatar_removal": [],
+                "banner_removal": [],
+                "federated_timeline_removal": [],
+                "followers_only": [],
+                "media_nsfw": [],
+                "media_removal": [],
+                "reject": [],
+                "reject_deletes": [],
+                "report_removal": []
+            },
+            "mrf_simple_info": {
+                "reject": {}
+            },
+            "quarantined_instances": [],
+            "quarantined_instances_info": {
+                "quarantined_instances": {}
+            }
+        },
+        "fieldsLimits": {
+            "maxFields": 10,
+            "maxRemoteFields": 20,
+            "nameLength": 512,
+            "valueLength": 2048
+        },
+        "invitesEnabled": false,
+        "localBubbleInstances": [],
+        "mailerEnabled": true,
+        "nodeDescription": "Akkoma: The cooler fediverse server",
+        "nodeName": "fedi.absturztau.be",
+        "pollLimits": {
+            "max_expiration": 31536000,
+            "max_option_chars": 200,
+            "max_options": 20,
+            "min_expiration": 0
+        },
+        "postFormats": [
+            "text/plain",
+            "text/html",
+            "text/markdown",
+            "text/bbcode",
+            "text/x.misskeymarkdown"
+        ],
+        "private": true,
+        "restrictedNicknames": [
+            ".well-known",
+            "~",
+            "about",
+            "activities",
+            "api",
+            "auth",
+            "check_password",
+            "dev",
+            "friend-requests",
+            "inbox",
+            "internal",
+            "main",
+            "media",
+            "nodeinfo",
+            "notice",
+            "oauth",
+            "objects",
+            "ostatus_subscribe",
+            "pleroma",
+            "proxy",
+            "push",
+            "registration",
+            "relay",
+            "settings",
+            "status",
+            "tag",
+            "user-search",
+            "user_exists",
+            "users",
+            "web",
+            "verify_credentials",
+            "update_credentials",
+            "relationships",
+            "search",
+            "confirmation_resend",
+            "mfa"
+        ],
+        "skipThreadContainment": true,
+        "staffAccounts": [
+            "https://fedi.absturztau.be/users/EpicKitty",
+            "https://fedi.absturztau.be/users/mizuki",
+            "https://fedi.absturztau.be/users/absturztaube",
+            "https://fedi.absturztau.be/users/enigmatico"
+        ],
+        "suggestions": {
+            "enabled": false
+        },
+        "uploadLimits": {
+            "avatar": 20000000,
+            "background": 40000000,
+            "banner": 40000000,
+            "general": 40000000
+        }
+    },
+    "openRegistrations": true,
+    "protocols": [
+        "activitypub"
+    ],
+    "services": {
+        "inbound": [],
+        "outbound": []
+    },
+    "software": {
+        "name": "akkoma",
+        "repository": "https://akkoma.dev/AkkomaGang/akkoma",
+        "version": "3.5.0-495-ga9eb94e8-fedi-absturztau-be"
+    },
+    "usage": {
+        "localPosts": 814773,
+        "users": {
+            "total": 1287
+        }
+    },
+    "version": "2.1"
+}
