+++ kitsune/Cargo.toml
@@ -58,12 +58,12 @@ rand = "0.8.5"
 rayon = "1.6.1"
 redis = "0.22.3"
 rsa = "0.8.1"
-sea-orm = { version = "0.10.7", features = [
+sea-orm = { version = "0.11.0", features = [
     "runtime-tokio-rustls",
     "sqlx-postgres",
 ] }
 serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.91"
+serde_json = "1.0.92"
 sha2 = "0.10.6"
 thiserror = "1.0.38"
 tokio = { version = "1.25.0", features = ["full"] }
@@ -78,7 +78,7 @@ zxcvbn = "2.2.1"
 
 [dev-dependencies]
 pretty_assertions = "1.3.0"
-sea-orm = { version = "0.10.7", features = ["sqlx-sqlite"] }
+sea-orm = { version = "0.11.0", features = ["sqlx-sqlite"] }
 
 [features]
 default = ["mastodon-api"]
